requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pymongo==4.6.1
redis==5.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.3
pyarrow==14.0.2

# Optional: Scalable Processing
# pyspark==3.5.0

# Image Processing & Forensics
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
Pillow==10.2.0
scikit-image==0.22.0

# Metadata Extraction
exifread==3.0.0
piexif==1.1.3

# Machine Learning
scikit-learn==1.4.0
scipy==1.11.4

# Apache Airflow
apache-airflow==2.8.0
apache-airflow-providers-postgres==5.9.0

# API & HTTP
httpx==0.26.0
requests==2.31.0
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Dashboard
streamlit==1.30.0
plotly==5.18.0
altair==5.2.0

# Report Generation
reportlab==4.0.9
fpdf2==2.7.7
matplotlib==3.8.2
seaborn==0.13.1

# Utilities
python-magic-bin==0.4.14  # For Windows
hashlib-additional==1.0.0
aiofiles==23.2.1
celery==5.3.4
flower==2.0.1

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.4
